src/python-backend/requirements.txt

streamlit>=1.22.0
google-generativeai>=0.3.0
